microservices/printer/requirements.txt
python-escpos==2.2.0
kafka-python==2.0.2
python-dotenv==1.0.0